Updated Scripts in Darwin on Vashishth.local

+++ macOS/update_scripts.sh
@@ -182,7 +182,9 @@ cp ~/.ssh/config .ssh/
 # Copy custom LaunchAgents
 mkdir -p LaunchAgents
 cp ~/Library/LaunchAgents/com.user.bluetooth.keepalive.plist LaunchAgents/
-cp ~/Library/LaunchAgents/com.user.codermana-nas-backup.plist LaunchAgents/
+if [ -f ~/Library/LaunchAgents/com.user.codermana-nas-backup.plist ]; then
+  cp ~/Library/LaunchAgents/com.user.codermana-nas-backup.plist LaunchAgents/
+fi
 
 echo "\033[1;31mRemoving bash unnecessary files...\033[0m"
 rm .bash_history
